refactor(cargo): Migrate to the dependencygraph API
Note that "Crate::autocfg:0.1.6" was dropped from expected results as it
was not referenced from any scope due to it being a build dependency of a
transitive dependency, and the dependency graph only reports referenced
packages.

Relates to issue #3825.
