+++ plugins/package-managers/cargo/src/funTest/kotlin/CargoSubcrateFunTest.kt
@@ -34,7 +34,7 @@ class CargoSubcrateFunTest : StringSpec({
         val definitionFile = getAssetFile("projects/synthetic/cargo-subcrate/Cargo.toml")
         val expectedResultFile = getAssetFile("projects/synthetic/cargo-subcrate-lib-expected-output.yml")
 
-        val result = CargoFactory.create().resolveSingleProject(definitionFile)
+        val result = CargoFactory.create().resolveSingleProject(definitionFile, resolveScopes = true)
 
         result.toYaml() should matchExpectedResult(expectedResultFile, definitionFile)
     }
@@ -43,7 +43,7 @@ class CargoSubcrateFunTest : StringSpec({
         val definitionFile = getAssetFile("projects/synthetic/cargo-subcrate/integration/Cargo.toml")
         val expectedResultFile = getAssetFile("projects/synthetic/cargo-subcrate-integration-expected-output.yml")
 
-        val result = CargoFactory.create().resolveSingleProject(definitionFile)
+        val result = CargoFactory.create().resolveSingleProject(definitionFile, resolveScopes = true)
 
         result.toYaml() should matchExpectedResult(expectedResultFile, definitionFile)
     }
@@ -52,7 +52,7 @@ class CargoSubcrateFunTest : StringSpec({
         val definitionFile = getAssetFile("projects/synthetic/cargo-subcrate/client/Cargo.toml")
         val expectedResultFile = getAssetFile("projects/synthetic/cargo-subcrate-client-expected-output.yml")
 
-        val result = CargoFactory.create().resolveSingleProject(definitionFile)
+        val result = CargoFactory.create().resolveSingleProject(definitionFile, resolveScopes = true)
 
         result.toYaml() should matchExpectedResult(expectedResultFile, definitionFile)
     }
